rename package as per standard
