+++ src/instance/sql_instance.go
@@ -4,7 +4,6 @@ package instance
 import (
 	"database/sql"
 	"fmt"
-
 	"github.com/newrelic/infra-integrations-sdk/v3/integration"
 	"github.com/newrelic/nri-mssql/src/connection"
 )
