+++ src/queryanalysis/connection/sql_connection_test.go
@@ -8,10 +8,12 @@ import (
 	"gopkg.in/DATA-DOG/go-sqlmock.v1"
 )
 
+var ErrMockCloseFailure = errors.New("mock close failure")
+
 func Test_SQLConnection_Close(t *testing.T) {
 	conn, mock := CreateMockSQL(t)
 
-	mock.ExpectClose().WillReturnError(errors.New("error"))
+	mock.ExpectClose().WillReturnError(ErrMockCloseFailure)
 	conn.Close()
 
 	if err := mock.ExpectationsWereMet(); err != nil {
