+++ src/queryanalysis/models/blocking_session_query_details.go
