+++ src/queryanalysis/models/customTypes.go
