+++ src/queryanalysis/models/database_details_dto.go
