+++ src/queryanalysis/models/query_details_dto.go
