+++ src/queryanalysis/models/top_n_slow_query_details.go
