+++ src/queryanalysis/models/wait_time_analysis.go
