+++ src/queryanalysis/validation/permissions_test.go
